feat(activate): Conditionally set JUPYTER_PATH
Prepend `JUPYTER_PATH` with `${FLOX_ENV}/share/jupyter` when that
directory is present, which happens when Jupyter related packages are
installed to an environment.

The `:+:` syntax to append with `:` if the `JUPYTER_PATH` is already
populated is a bit awkward to read but arguably less clunky than testing
the variable and exporting either side of an if statement, plus we
already use the same syntax elsewhere.

Nesting environments turned out to be a good way to also verify that
prepending behaviour. The test cases are combined together because they
share the same setup costs.

This functionality depends on our change to the upstream package in:

- https://github.com/NixOS/nixpkgs/pull/341093

However since that package is only "versioned" by the version of Python
that it builds against we can't provide Flox users with more precise
instructions than to `flox upgrade` to the latest and hope that one of
their existing packages doesn't hold the base page back too far.