test: split up activation tests
Some of our activation tests check multiple behaviors in the same test,
and they check that `hook.on-activate` is run multiple times. Since
we're about to add support to attach to an existing activation, if a
second `flox activate` attaches, it won't run `hook.on-activate` a
second time, and the test will fail.

We could wait for the first `flox activate` to get cleaned up, but since
in general we prefer more targeted tests, just split the tests up
instead.