fix(services-tail): stop on final line on stderr
`process-compose logs --tail` immediately prints logs
and then takes about a second to eventually exit.
In order to avoid this unnecessary pause, we utilise the fact
that process compose will print a warning message

    write close: write unix ->/tmp/.tmpnYlCZ2/S.process-compose: write: broken pipe

as the last line.
Previously, this warning was (incorrectly) printed to stdout as well,
which was eventually fixed upstream.
However, since we inspected only the stdout stream, we were not skipping the pause anymore,
which in effect broke the 'logs: tail does not wait' integration test.

The proposed fix is to combine the output streams of process compose and use them,
as if all messages were printed on a single stream, simulating the previous behaviour.

Eventually, we want to avoid trying to stop reading early, and instead have `process-compose`
exit as soon as possible.