feat: selectable activation-scripts "interpreter" package
Going forward we will refer to the files in the flox-activation-scripts
package as the "flox interpreter", because it acts like an interpreter
as it "activates" the contents of a $FLOX_ENV. It is also like an
interpreter in the classic UNIX sense because it can be both included
with the environment (just like the "shebang" statement in a script)
or used from outside a previously-rendered environment to provide
updated functionality without having to re-render the environment
(just like you can run a script with `/path/to/interp <script>`).

This patch adds the ability to select at runtime whether to activate
environments with:

* the latest interpreter as bundled with the flox CLI (default), or
* the interpreter installed to the environment as it was built

By offering both interpreters to the user at activation time we get
the maximum flexibility, providing the ability to move forward with
updated activation features over time by default while preserving the
ability to revert to the "original" features as installed to the
environment at build time. This patch introduces the new
`--use-old-interpreter` flag for controlling this behavior.