fix: set runtime dir for pure builds
Pure builds run in a sandbox, so flox-activations will currently error
with:
```
[2024-10-24T17:26:13Z DEBUG flox_activations::cli::start_or_attach] Reading activations from "/homeless-shelter/.cache/flox/721f90fc/activations.json"
Error: failed to acquire lockfile
```
(that includes output from a debug! added in this commit)

Introduce FLOX_RUNTIME_DIR which the pure manifest builder can pass to
the activation script and then flox-activations. Set it to TMP, since
this will be set to something writeable in the sandbox.

This approach is consistent with the approach suggested for
FLOX_RUNTIME_DIR in
https://github.com/flox/flox/issues/2240, but it doesn't yet make all
those changes.