refactor(catalog): Move reset mocks to test_helper
So that we don't need to repeat the `if panic` dance everywhere.

Keep the underlying implementation on `MockClient`, so that it's close
to the dependent functions, but make it private so that we're forced to
use the test helper.

Using the `test_helpers` mod rather than `impl Client` so that we don't
accidentally use it outside tests, which would only panic at runtime.