refactor(watchdog): Only use --pid for register
Only use the `--pid` argument for registering an activation PID in the
environment registry. We don't need to pre-populate the list with the
original PID in `wait_for_termination` because `update_watchlist` will
read the already registered activation.

After switching to `activations.json` we'll no longer need the PID
argument because `start-or-attach` will have already registered it.