fix: set LC_ALL=C for sort and comm
On Linux, running
```
LANG= flox-cli-tests -- -f "can start redis-server and access it using redis-cli"
```
fails with
```
   /nix/store/0kg70swgpg45ipcz3pr2siidq9fn6d77-coreutils-9.5/bin/comm: file 1 is not in sorted order
   /nix/store/0kg70swgpg45ipcz3pr2siidq9fn6d77-coreutils-9.5/bin/comm: input is not in sorted order
```

This is because in our activation scripts, we:
- Call sort
- Source the activation hook
- Call comm

The redis activation hook sets LANG, which causes sort order to change
between when we call sort and when we call comm, so comm fails.

Set LC_ALL=C for all sort and comm commands since we don't really care
about the ordering and just need something consistent.