refactor(activations): Rename flox_env help
This argument is passed `$FLOX_ENV` which is `.flox/run/{arch}.tmp` and
a symlink to the environment's store path rather than `.flox` itself.

I considered whether we could use this to calculate `id` and
`store_path` in some places, rather than passing explicit args, but
there's the potential that the symlink could change underneath us while
an activation is starting.