refactor(watchdog): Move watchdog to start.bash
Start the watchdog from `start.bash` just before `set-ready` is called
so that it's only started once for the first "start" activation and not
the subsequent "attach" activations.

I previously wanted to do this within `flox-activations set-ready` so
that we got some type safety on the command/argument handling and it
would be atomic with the `set-ready` operation. However the need to
make this conditional for direct invocations of `${FLOX_ENV}/activate`,
where we don't currently want to start a watchdog and can't easily
compute all of the arguments (e.g. `--log-dir` when we don't have
`.flox` path), later convinced me to just implement it in Bash.

Passing the path to the watchdog binary in at runtime also ensures that
the watchdog isn't unnecessarily bundled into the closure of
environments, including containers which don't need a watchdog because
there is always a single "start" activation for the lifetime of the
container.

The `--pid`, `--registry`, and `--hash` arguments can be removed when
the watchdog is subsequently changed  from using the environment
registry to `activations.json`.

We no longer explicitly skip starting a watchdog for "ephemeral"
activations, where we call `${FLOX_ENV}/activate` for the purpose of
starting `process-compose` from the right environment, mostly because I
don't want this concept to escape the CLI with yet another environment
variable. This is fine though because it will go through the normal
`flox-activations` flow, starting or attaching, and we are going to
configure the watchdog to only cleanup `process-compose` when _all_
activations for a `$FLOX_ENV` have exited.

We also no longer skip for "in-place" activations but they go through
the normal `flox-activations` flow and we intend to support services for
them very shortly.

There are two outstanding issues to still address:

1. Wrapped binaries from `flox build` will "start" an activation without
   starting a watchdog, which could prevent cleanup or hooks being run
   whilst a wrapped binary is being run. This shouldn't be a release
   blocker for the activation work because build isn't generally
   available yet so we will followup in #2287

2. We have a new race condition, which I noticed by chance, where
    `process-compose` isn't being cleaned up for for multiple `flox
    services [re]start` commands.

   The watchdog removes the activation from `env-registry.json` and
   stops `process-compose` on the first iteration of the loop but
   nothing removes the activation from `activations.json`, so the second
   iteration of the loop attaches to an existing activation and doesn't
   start a watchdog.

   This will be fixed by #2130 and #2286

    ```
    % just integ-tests services.bats -- --filter-tags services:permit-stopped-for-start-restart
    …
    services.bats
     ✓ start, restart: do not require active services [11703]
    Waiting for pids: 33511

    1 test, 0 failures in 13 seconds

    % pstree -ws bats
    -+= 00001 root /sbin/launchd
     \-+= 33600 dcarley /nix/store/dxgk503lw2a0slqcvhcvwfa07qf9y8sx-process-compose-1.27.0/bin/process-compose up -f /private/var/folders/8q/spckhr654cv4xrcv0fxsrlvc0000gn/T/nix-shell.9nd2Vl/bats-run-DlcaDa/test/1/test/.flox/run/aarch64-darwin.test/service-config.yaml -u /private/tmp/flox.tests.LgLx2p/tmp.x68yCqYIQZ/run/flox.825e1522.sock -L /private/var/folders/8q/spckhr654cv4xrcv0fxsrlvc0000gn/T/nix-shell.9nd2Vl/bats-run-DlcaDa/test/1/test/.flox/log/services.20241029104728921926.log --tui=false
       |--= 33605 dcarley /nix/store/1l41471x6rlf7l544s1bxkqqgpprd28m-coreutils-9.5/bin/sleep infinity
       |--= 33606 dcarley sleep infinity
       \--= 33607 dcarley sleep infinity
    ```