feat: update activation scripts on new version
This makes a change to how activation scripts are stored and used. Now
every flox CLI will bundle its own activation scripts that it will use
to activate environments instead of the ones stored in the rendered
environment. New environment builds will store the activation scripts
that are bundled with the CLI.

This behavior serves two purposes:
- It allows the latest activation scripts to be used without rebuilding
  the environment.
- It allows you to fall back to the activation scripts that were current
  at the time the environment was built.

This also introduces a naming change. The activation scripts are now
referred to as the `flox-interpreter` since they "interpret" and execute
the rendered environment.