refactor: simplify update_watchlist logic
Our current update_watchlist logic calls 3 different functions which all
check if a PID is still running:
- activation_pids() calls remove_stale_activations()
- prune_terminations()
- watch_new_pids()

The logic to determine if a PID is about to get more complicated (since
we'll need to check if an attached PID has expired), so it's preferable
to have to update just one thing instead of 3.

Change the algorithm in update_watchlist() such that PIDs are checked to
be running just once.