build: bundle `nixpkgs` used at build time
`flox build` should use a known and bundled version of nixpkgs, so that
a) build issues related to the nixpkgs used do not depend
   on the user's environment or change from one environment to the other.
b) builds do not require downloading and unpacking nixpkgs at runtime.

With this change, `flox build` now depends on the same **store path**
of nixpkgs as the build of flox itself,
i.e. the `nixpkgs` input of the flox crate.

```
  % nix eval --impure
  % nix build .#flox
  % nix path-info --recursive .#flox
  ...
  /nix/store/p2hby44a0qzrnd1vxcpcgfav6160rmcv-source

  % flox build
  ...
  nix build [..] --argstr nixpkgs-url path:/nix/store/p2hby44a0qzrnd1vxcpcgfav6160rmcv-source
```

In addition we updated the pkgdb/default.nix to use the same `path:` url,
and make both implementations more similar to each other.