chore(watchdog): Move from bin to libexec
So that it doesn't get installed to `/usr/local/bin` on MacOS and end up
leaking onto `$PATH`. We don't want to expose this to users because they
never need to interface with it directly.

My understanding is that `flox-installers` collects up these packages
and puts them into `/usr/local`, so that moving it within these packages
will have the desired outcome.