refactor: set FLOX_RUNTIME_DIR in CLI
flox-watchdog is going to need runtime_dir. We currently compute
runtime_dir in flox-activations, but per
https://github.com/flox/flox/issues/2240 we'd like to factor that out.
This is because runtime_dir may be different for each of
`flox activate`, `flox build`, and `flox containerize`, so each of those
should be able to pass runtime_dir to the activation script.

This commit passes runtime_dir through with an environment variable from
the CLI.

There are a few things I don't like about the current approach, but I
left as is for the sake of time:
- runtime_dir is added to Flox but not Config. That's because it's
  sometimes derived from cache_dir, and it seems hard to make a config
  field depend on another config field. I don't really think we need it
  in Config, but I still don't like the asymmetry
- It seems like it would be preferable to make --runtime-dir a required
  arg to the activation script rather than failing when FLOX_RUNTIME_DIR
  is not set. And it's a bit confusing that it gets set in the CLI even
  though the call to the activate script happens in package-builder.
- We shouldn't have to set FLOX_RUNTIME_DIR for binaries we're wrapping.
  I think there will be changes soon to how we wrap binaries anyways, so
  I just left a TODO for now.