fix: local test failure
When running the test `kills daemon process` locally it fails. This is
because flox-cli-tests sets FLOX_WATCHDOG to `flox-watchdog` instead of
an absolute path. Our daemonize call in the activation scripts then
fails with `The 'path' parameter must be an absolute path name.`

Use `which flox-watchdog` to set an absolute path instead.