feat(containerize): wait for subprocesses
fix: wait for container stream process

fix: wait for child container sink process
Ensure that the container image has **successfully** been written to the sink.
In case of files that is done by fsync'ing the file.
When piping into a runtime directly,
we wait for the runtime process and report a failure if the process exited unexpectedly.