feat: prompt to create default environment
This prompts the user to create a default environment the first time
that they attempt to install a package without an activated environment
and without an environment in the current directory. This choice is
persisted in a new file called `user_state.json`. The field for this
choice is an Option<bool> so we can represent "unprompted" as None, and
the user's choice as Some(value). If the user cancels the operation, no
choice is persisted.