refactor: build environment using `pkgdb realise` and `nix build`
Reimplement `CoreEnvironment::build` using `pkgdb realise`
and `nix build --file <buildenv.nix> [...]`.

At this point the goal is tho have the implementation in place
while making the smallest amount of changes necessary.
That also means that optimizations such as calling realise for missing store paths only,
and introducing dedicated error messages for the `nix build` step
are omitted for expediency.

In the spirit of our other external subsystems,
the implementation is likely to wander off into a dedicated `provider` module
in a follow up change.

fix: update unit test for renamed develop environment

The realpath to "bin/hello" has changed with the renaming of develop
environment to have a "-develop" suffix.