fix: bugs in buildenv.nix
fix: render profile-common script from manifest

Missed this bug in new buildenv implementation.

fix: buildenv was copying service-config.yaml to wrong place

Although ... it's arguable the root directory isn't the right place either,
but that's where it's currently expected!

fix: do not access null values in buildenv.nix

Somehow the interaction of a Nix 2.18.7 client and 2.24.6 nix-daemon triggered
errors attempting to expand null values that didn't happen on systems running
a 2.18.7 nix-daemon. Perhaps it's a change in the Nix evaluator, but regardless
updating buildenv.nix to make doubly certain not to access null values found
in the manifest.lock file.