feat: auto add activation for default env
This prompts the user whether they would like us to add an automatic
activation of the default environment to their shell's RC file. We look
in the standard locations for the RC files, creating one if it doesn't
exist, making a backup before writing to it, and finally append an
in-place activation line to their RC file.