feat(CommandExt): include env changes (if any)
Using `DisplayCommand` can be used to print a copy-pastable shell command.
So far this only included arguments.
However some commands _also_ set or unset environment variables,
which is currently not visible in the logs.
This commit makes `DisplayCommand` print an invocation in the form of

    [env (-u VAR | VAR=VALUE) ... ] CMD ARGS...

if either `Command::{env,envs}`,  or `Command::env_remove` was used on the printed command.