refactor: move `buildenv` implementation behind a provider trait
In the spirit of consistency with isolating our subsystems --
i.e. external service providers like services, git and (manifest) builds --
move the logic of calling the constituents of building
and linking environments to a dedicated `buildenv` provider.

Ideally `link`/`linkenv` would be a separate trait and separate implementation,
but at the moment its included in the `BuildEnv` trait out of practicality

* linking is trivial
* it was originally implemented by the same tool (`pkgdb {link,build}env`)
* linking is currently only performed for environments