fix: install "outputsToInstall" packages with higher priority
The previous commit (9ae08258) which disabled outputsToInstall support had
the effect of installing all outputs with the same priority, which then
introduced the potential for collisions when multiple outputs contained
the same paths with differing contents. One example of such a package
is the most recent postgresql, as detailed in #2336.

This patch revises the previous commit by installing "outputsToInstall"
outputs separately from other outputs, and tweaking their respective
priorities so that we don't end up with collisions between any of the
"other" outputs that we add by default.