feat: add `RenderedEnvironmentLink`
A link to a built environment in the Nix store.

The path may not exist if the environment has never been built and linked.

As part of an environment, the existence of this path guarantees exactly two things:
- The environment was built at some point in the past.
- The environment can be activated.

The existence of this path explicitly _does not_ guarantee
That the current state of the environment is "buildable".
The environment may have been modified since it was last built
and therefore may no longer build.

feat: add `RenderedEnvironmentLinks` constructor