refactor: replace `out_path` with `rendered_env_links` in remote env
feat: migrate gcroots of remote environments

Remote environments use a ManagedEnvironment to manage the environment.
Previously we linked a gcroot to `<path of managed environment>/run`.
Note, that this was a single link.
For consistency with the new more consistent naming scheme,
we want to create links in a `<path of managed environment>/run` __directory__!

As such we need to migrate the single link by deleting it,
so we can create a directory in the step following.