feat: wrap manifest builds with build-specific env
Following the recent buildenv replacement we now build dedicated closures
for each manifest build package. This patch updates the builder to wrap
files in bin and sbin with this new closure instead of the original
$FLOX_ENV closure used to perform the actual build.

Note that it's not just enough to wrap the built executables with the "build"
environment, but we must also explicitly compile the package to prefer it
over the "develop" environment for library paths, interpreters, etc.
This also patch updates the build invocations to activate the "develop"
env in runtime mode, and then the "build" env in develop mode, so that
the resulting build invocations prefer libraries and tools found in the
"build" env.