chore: fix typos and style
This is a rebase of 468a2cd from the contributor ccoVeille against
`origin/main` at 30ec35c6 because we delayed merging it while having
problems with CI and I didn't want to ask the contributor rebase again
several times.

It fixes merge conflicts and omits `catalog-api-v1/src/client.rs`
because it's generated code from another repo. I'll raise a separate PR
for that.

I've reviewed all of the changes again, so I just need a stamp to merge
and then I'll close 2362 with an update.