feat(activate): Warn on incompatible activations
Add a mechanism for requiring users to exit existing activations when
using an incompatible version of `flox-activations`. We can bump this
version when making breaking changes to activations, such as introducing
new environment variables between the CLI and interpreter.

By implementing this in `flox-activations` it also prevents services
from being started because they use an ephemeral activation which goes
through the same code path.

It's worth noting that because the version is applied to the root of
`activations.json` that it will catch all activations of all store paths
for an environment. It will not however catch different runmodes of an
environment because they use different `activations.json`, per 2357.

We are not going to add any checks for old non-attach style activations
because we've already shipped that change in v1.3.4 and it would add an
unwanted dependency on the environment registry to `flox-activations`
and `flox-watchdog`.

The error message differs slightly from the original acceptance criteria
in that:

- we don't have access to the environment name without introducing an
  implicit dependency on `FLOX_ENV_DESCRIPTION` but most of our other
  error messages don't include it so I think it's fine to omit

- a list of activation PIDs is included to indicate to the user which
  processes they need to exit, since it was easy to get

Things that I feel uneasy about:

- Inconsistent error formatting; `flox-activations` is getting wrapped
  by `activate` and sometimes `flox activate` so we have limited control
  over how the final error message is formatted, as shown by the tests.

- Using `$PPID` rather than `$$` in the test; possibly because the
  command activation gets wrapped with `FLOX_SHELL` first?

- Maintaining two very similar tests for `activate` and `services`; they
  are different enough and they live in the correct respective places
  but it still sucks to duplicate.