chore(activate): Reinstate -euo pipefail
So that errors from `flox-activations` get surfaced early and prevent
`flox activate` from proceeding in an unknown state.

This is required for an `activations.json` version check that will be
added in a subsequent commit. It would have also caught this previous
issue with downgrading `flox` against an updated environment:

    % nix run 'github:flox/flox/v1.3.3' -- activate -d ~/tmp -- echo hello
    error: a value is required for '--runtime-dir <PATH>' but none was supplied

    For more information, try '--help'.
    /nix/store/rx3gpsqb5iql6azx9fdp1b31070h70q1-flox-activation-scripts/activate.d/start.bash: line 14: /bare.env: Read-only file system
    /nix/store/rx3gpsqb5iql6azx9fdp1b31070h70q1-flox-activation-scripts/activate.d/start.bash: line 39: /post-etc-profiles.env: Read-only file system

This is a partial revert of 0110e0d4 now that the CUDA issue has been
fixed in a preceeding commit.