fix(bats): Unset FLOX_ENV* and _FLOX_ENV
Unset any `FLOX_ENV*` and `_FLOX_ENV*` environment variables at the very
start of the suite to prevent an outer activation from leaking into the
tests.

I use a Flox activation as my default shell and this was surfacing two
kinds of test failures for me.

Primarily all of the "running activate directly" tests were failing with
missing `sourcing hook.on-activate` since we introduced the start/attach
behaviour. I believe this is because they were attaching to my outer
activation when bypassing `flox activate`. Full output here:

- https://gist.github.com/dcarley/c141e064cdbd33f03dcba1eb41cd0d52

Secondarily the PATH of my outer activation was leaking test
activations. I noticed this when trying to use `yq` for something, which
has a different version in my default environment vs the `devShell`. It
can be observed with this test:

    @test "wip: yq should come from devShell" {
      project_setup

      EXPECTED_PATHS="$(which -a yq)"
      run "$FLOX_BIN" activate -- which -a yq
      assert_output "${EXPECTED_PATHS}"
    }

Which would fail with:

    -- output differs --
    expected (1 lines):
      /nix/store/x2hwfhcs7mcd3sslvkvraw32l3wjksw2-python3.12-yq-3.4.3/bin/yq
    actual (4 lines):
      /Users/dcarley/projects/dcarley/flox-envs/shell/.flox/run/aarch64-darwin.shell.run/bin/yq
      /nix/store/x2hwfhcs7mcd3sslvkvraw32l3wjksw2-python3.12-yq-3.4.3/bin/yq
    --

After this change all of those tests pass. There may be more that we
need to unset for the interpreter to be truly isolated but we can't
unset everything `FLOX*` or `_FLOX*` because `flox-cli-tests` uses it to
pass some information over to BATS. This should be sufficient for the
moment.