feat: define verbosity macros based on $_FLOX_PKGDB_VERBOSITY [sic]
It is really useful to be able to tune makefile verbosity in response to the
`flox -v` verbose flag, and this commit introduces 3 new macros for this
purpose, defined as in this pseudo-code:
```
  _V_ = $_FLOX_PKGDB_VERBOSITY > 0 ? "" : "@"
 _VV_ = $_FLOX_PKGDB_VERBOSITY > 1 ? "" : "@"
_VVV_ = $_FLOX_PKGDB_VERBOSITY > 2 ? "" : "@"
```
The end result is that we can preface any line in a makefile stanza with
`$(_V_)` and make will only report that invocation if flox has been invoked
with `-v`. Similarly a line starting with `$(_VVV_)` will only be reported
if flox has been invoked with `-vvv` (or more).