test(activate): Workaround bashrc_Apple_Terminal
Running the tests in `Terminal.app` on Darwin caused this failure
because it loads customisations from `/etc/bashrc_Apple_Terminal` that
affect the output:

    -- values do not equal --
    expected : 7
    actual   : 9
    --

    Last output:
    Sourcing .profile
    Setting PATH from .profile
    sourcing hook.on-activate
    Sourcing .bashrc
    Setting PATH from .bashrc
    sourcing profile.common
    sourcing profile.bash

    Saving session...completed.
    Deleting expired sessions...none found.

Disable it by the means suggested in that file. I tried setting
`SHELL_SESSION_HISTORY=0` but that didn't change this behaviour.

I don't particularly care for `Terminal.app` but I occasionally reach
for it instead of `Kitty.app` when I want to bypass my default Flox
environment because it's causing problems like #2369.