fix: propagate error from process substitution
The source tarball target was employing `$(git ls-files)` to get the
source manifest, but when this failed the invocation would successfully
generate an empty tarball rather than failing. TIL that you need to
follow a process substitution with a call to `wait $!` in order to
get the desired behaviour!

Having fixed the bug, also needed to an invocation of GitCommandProvider::init
to the rust build tests so that it doesn't crash when performing a sandbox
build from a git-less directory.