fix(publish): use local HEAD for publish rev
Current code uses the origin info to get the rev to publish, but this is
the latest in the remote, which may not be avalible locally.  We want to
publish the local current rev, since that's what was built.  Later we
will add checks that this commit also exists in the remote, but that is
a check for later.  This fix allows you to publish from your local clone
without being required to have the latest pulled down.