refactor(activations): Guard version states
Use types to indicate whether the version of `activations.json` has been
checked or not. Methods that don't need to check the version have been
moved out to a separate trait. Most notably used by the watchdog because
it should never exit prematurely when it's backgrounded and has no way
to recover or signal to the user.

This also fixes a bug whereby starting an activation for a new store
path would succeed but attaching to it would fail. We now check the
version for both start and attach.