fix(tests): Default SHELL and unset FLOX_SHELL
Use `bashInteractive` from `flox-cli-tests` rather than leaking `SHELL`
from the user running the tests.  This surprised me when writing a `flox
activate` test that made use of `export -f` because it was defaulting to
my `zsh` shell.

Without the change to `setup_suite.bash` the new test fails for me with:

    ✗ suite: SHELL should default to bashInteractive from Nix [30]
       tags: suite
       (from function `assert_output' in file /nix/store/h7inz07x6739s38i02532g5pj6dlnx8k-bats-with-libraries-1.11.0/share/bats/bats-assert/src/assert_output.bash, line 178,
        in test file suite.bats, line 19)
         `assert_output --regexp "^/nix/store/.*-bash-interactive-.*/bin/bash$"' failed
       /bin/zsh

       -- regular expression does not match output --
       regexp : ^/nix/store/.*-bash-interactive-.*/bin/bash$
       output : /bin/zsh
       --

       Last output:
       /bin/zsh