test: increase timeout of service log tests that follow
Previously the integration tests tagged with `services:logs:follow:combines`
expected to see output after 0.5 to 1 seconds
after calling `flox services logs -f`.

It was expected that that is enough
because the flox command was run after we knew,
that the processes have printed output,
by way of waiting for a pipe signal.
In addition, we expect all a tail of the previous output
to be printed immediately.

Unfortunately "immediately" appears to take longer than 0.5 to 1 seconds,
depending on the test.
To combat this we now increase the timeout to first print,
to up to 5 seconds.

Increasing a timeout is not ideal (timeouts in general aren't),
however we still need to stop reading logs at some point,
as `--follow` will keep stdout open virtually forever.
We could remove the timeout altogether
and poll the output file for the expected output, blocking forever,
if no output is received.
Especially in CI a non finishing test has been found
to be hard to detect, and attributed to different things
(e.g  leaking fd3).
To avoid this, an excessively long timeout --
we already know the processes print and waiting for that message
to make it through the pipeline -- will at least help us to see
process-compose communication issues earlier.

It be noted, that the unit tests _do not_ apply any timeout,
but wait for output, which they eventually receive.