fix: pass compiler-related variables to "inner" activation
When building in situ, compilers found in the outer "develop" activation
need to communicate certain environment variables to their invocations
as performed in the inner "build" activation. This introduces a real
chance that references to the "develop" environment will leak into the
build package, but the alternative is that the compilers will fail
altogether, so this may be one of those instances where manifest builds
are less capable than their pure NEF counterparts.