feat: reduce log output
The watchdog was emitting log lines at a rate of 10Hz in some cases,
which quickly generates enormous log files. This reduces the log level
of the two offending lines and allows a user to override the log level
of the watchdog (mostly for testing purposes).

This also adds log rotation to the log files created by the watchdog
specifically. The logs from services are still put into a single file by
process-compose. New log files are created daily for a given activation,
so a long-running activation will only ever have ~3 watchdog log files
on disk given that we garbage collect log files that haven't been
modified in at least 3 days. The log files have names of the form
'watchdog.{hash}.log.{yyyy-mm-dd}'.

The GC of log files also now runs in a loop in a background thread. This
background thread is only ever cleaned up when the watchdog dies, just
like the heartbeat thread, so you can't really put it in a test.
Instead, the logic for detecting which logs to GC was factored out and
tested in isolation (there's not much value in testing that `rm` works).