chore(push): Generalise FloxHub edit message
State that new environments are public by default and give a URL where
they can navigate to settings (as well as promoting other view/edit
options) whilst still giving us us the flexibility to change in FloxHub
which environments can be made private, both now and again in future.

I briefly looked at converting these to use `env.floxhub_url` but then
got lost in a rabbit hole of what to print when `join()` fails and
whether we should promote the URL for all messages, so I've deferred
that for now.