chore: tweak manifest build verbosity
Following the addition of the makefile verbosity macros, use them to
improve the signal-to-noise ratio of the text coming out of manifest
builds. This is just an initial stab to get the ball rolling; I don't
doubt there will be further modifications to follow.

This commit updates the CLI to thread the verbosity value through to
the build module.  It also sets verbosity to 1 in rust tests that are
looking for lines that are now only included in the build output when
flox is invoked with `-v`.