feat: allow overriding opened browser with `BROWSER`
To test our login flow I need to be able to specify the browser, that is being used.
The only way to so this now is to change the system default browser
(because flox internally uses `open` -- or linux equivalents i available -- which is hard wired to use the system default browser).
Which means my default browser gets reset if I run the tests locally.
One thing that is quite common with web dev tools, is using the BROWSER environment variable to specify a custom browser.

This commit adds support for browser detection via `BROWSER`, and performs light argument parsing,
such that `BROWSER='firefox -P some-profile' can be set to use a specific browser profile for example.