fix(remote): Honour trust for case sensitive envs
The `config` crate introduced a regression in 0.14.0 (which we upgraded
to in 6588ba72) whereby all keys are lowercased. This isn't configurable
And doesn't look like it will be "fixed" any time soon, so we're
downgrading to 0.13.4 which was the last release before the change.

The removal of unused features in the preceding commit makes this pretty
simple and low risk in terms of introducing CVEs from downgraded
transient dependencies.

This prevented us from honouring saved trust choices for case sensitive
environments like `rossturk/ComfyUI` because the choice was saved as
`ComfyUI` but read back as `comfyui`.

Includes a test that reproduces and prevents regressions of the issue. I
looked at adding a more targetted unit test but it's not especially easy
to parse a `Config` from a specific location and the integration test
clearly demonstrates the user effect. Prior to the downgrade the test
would fail with:

    ❌ ERROR: Environment owner/CaseSensitive is not trusted.

    flox environments do not run in a sandbox.
    Activation hooks can run arbitrary code on your machine.
    Thus, environments need to be trusted to be activated.