test(services): Refactor nested comments
The single quotes within comments of EOF blocks having been causing
problems with syntax highlighting in my editor and I've given up finding
a solution that doesn't impose this on others.

Replaced the comments about not using `set -euo pipefail` with explicit
`|| true` to make it clearer that we don't care about the exit codes.