test: fix systems for pull integration test
The flox pull test pulls an environment created by the flox push test in
a previous run of the integration tests. Both tests are only run on
x86_64-linux and aarch64-darwin, but we're currently trying to pull an
x86_64-darwin environment, which doesn't exist.

I'm not completely sure how an x86_64-darwin environment existed before,
but we may have run the flox push test manually on an x86_64-darwin
machine or on a different runner in the past. Since we just changed the
FloxHub user we're testing as, it surfaced the issue.