fix: don't always rebuild remote environments
Currently every `flox activate -r` triggers a rebuild of the
environment, which is slow.

This is happening because RemoteEnvironment::new_in runs a force pull of
the underlying ManagedEnvironment.

There's currently a logic error in pull where force pulls don't check if
an environment is up to date, so a force pull returns
PullResult::Updated even when it should return PullResult::UpToDate.

Coincidentally, this leads to calling
reset_local_env_to_current_generation(), which changes the mtime of the
manifest, which causes ManagedEnvironment::rendered_env_links to call
build().

To fix this, check if the environment is up to date for force pulls in
the same way as normal pulls.